Fix compilation on FreeBSD

+++ src/runner.cpp
@@ -16,7 +16,7 @@ RemoteImage Runner::serializeImage(const QImage &image)
     RemoteImage remoteImage{
         convertedImage.width(),
         convertedImage.height(),
-        convertedImage.bytesPerLine(),
+        static_cast<int>(convertedImage.bytesPerLine()),
         true, // hasAlpha
         8, // bitsPerSample
         4, // channels
